requirements.txt

biopython==1.81
biotite==0.36.1
requests==2.28.1
et-xmlfile==1.1.0
fair-esm==1.0.3


networkx==3.1

openpyxl==3.1.2

pandas==1.4.4
scipy==1.9.3
seaborn==0.12.2
torch==1.13.0

torchaudio==0.13.0
torchvision==0.14.0
torch-geometric==2.1.0

scikit-learn==1.1.3

